requirements.txt
fastapi 
uvicorn 
langchain 
openai 
faiss-cpu 
pydantic 
PyPDF2
python-dotenv
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
google-cloud-speech
google-cloud-texttospeech
fitz
textract